Prefer the last modification time of the cache objects
The cache objects can be newly created or updated.  When the cache
object is updated by a build, it makes more sense to use the last
modification as an 'update time' instead of the creation time.

This patch factors out a `getCreateTime` function which tries to
retrieve the last modification time and the object creation time, in
this order.